Make offsetFromReq and limitFromReq reusable
