+++ api/helpers/things.js
@@ -13,7 +13,7 @@ let { remove } = require("lodash");
 const cache = require("apicache");
 const equals = require("deep-equal");
 const moment = require("moment");
-const { SUPPORTED_LANGUAGES } = require("./../../constants.js");
+const { SUPPORTED_LANGUAGES, RESPONSE_LIMIT } = require("./../../constants.js");
 const logError = require("./log-error.js");
 const createCSVDataDump = require("./create-csv-data-dump.js");
 
@@ -158,6 +158,24 @@ const uniq = list => {
     }
   });
   return newList;
+};
+
+const limitFromReq = req => {
+  let limit = parseInt(req.query.limit || RESPONSE_LIMIT);
+  const resultType = (req.query.resultType || "").toLowerCase();
+  const returns = (req.query.returns || "").toLowerCase();
+  if (resultType === "map") {
+    limit = 0; // return all
+  } else if(returns === "csv") {
+    limit = 0;
+  }
+  return limit;
+};
+
+const offsetFromReq = req => {
+  let query = req.query.page ? req.query.page.replace(/[^0-9]/g, "") : "";
+  const page = Math.max(as.number(query || 1), 1);
+  return (page - 1) * limitFromReq(req);
 };
 
 let queries = {
@@ -497,5 +515,7 @@ module.exports = {
   placeHolderPhotos,
   createLocalizedRecord,
   getCollections,
-  validateFields
+  validateFields,
+  limitFromReq,
+  offsetFromReq
 };
